fixed bug arrising due to type errors

+++ macrosynergy/management/check_availability.py
@@ -107,7 +107,10 @@ def visual_paneldates(df: pd.DataFrame, size: Tuple[float] = None):
         # in the DataFrame.
         maxdate = df.max().max()
         df = business_day_dif(df=df, maxdate=maxdate)
-        df = df.astype(int)
+
+        df = df.astype(float)
+        # Ideally the data type should be int, but Pandas cannot represent NaN as int.
+        # -- https://pandas.pydata.org/pandas-docs/stable/user_guide/gotchas.html#support-for-integer-na
 
         header = f"Missing days prior to {maxdate.strftime('%Y-%m-%d')}"
